Fix bug where table headers were appearing greyed out. Add null safety checks to Account Management tab. Refactor layout for Account Management tab.

+++ immigration-analytics/src/main/java/com/nineplusten/app/view/AccountManagementController.java
@@ -3,16 +3,13 @@ package com.nineplusten.app.view;
 import com.nineplusten.app.service.LoadUserDataService;
 import com.nineplusten.app.service.DeleteUserService;
 import com.nineplusten.app.model.UserData;
-
 import javafx.fxml.FXML;
 import javafx.scene.control.TableView;
 import javafx.concurrent.WorkerStateEvent;
 import javafx.event.EventHandler;
-
 import java.util.List;
 import javafx.event.ActionEvent;
 import java.util.ArrayList;
-
 import javafx.collections.ObservableList;
 import javafx.collections.FXCollections;
 import javafx.scene.control.TableColumn;
@@ -34,6 +31,8 @@ public class AccountManagementController {
   private TableColumn<UserData, String> roleColumn;
   @FXML
   private Button refreshButton;
+  @FXML
+  private Button deleteButton;
 
   LoadUserDataService userService;
 
@@ -41,29 +40,34 @@ public class AccountManagementController {
 
   @FXML
   private void refreshPage(ActionEvent event) {
-    userService.restart();
+    if (!userService.isRunning()) {
+      userService.restart();
+    }
   }
 
   @FXML
   private void deleteUser(ActionEvent event) {
-    String userID = userTable.getSelectionModel().getSelectedItem().userIDProperty().getValue();
+    UserData selectedItem = userTable.getSelectionModel().getSelectedItem();
+    if (selectedItem != null) {
+      String userID = selectedItem.userIDProperty().getValue();
 
-    Alert alert = new Alert(AlertType.CONFIRMATION);
-    alert.setTitle("Confirm");
-    alert.setHeaderText("Deleting User");
-    alert.setContentText("Are you sure you want to delete the user: " + userID + "?");
+      Alert alert = new Alert(AlertType.CONFIRMATION);
+      alert.setTitle("Confirm");
+      alert.setHeaderText("Deleting User");
+      alert.setContentText("Are you sure you want to delete the user: " + userID + "?");
 
-    alert.showAndWait();
+      alert.showAndWait();
 
-    if (alert.getResult().getText().equals("OK")) {
-      deleteService = new DeleteUserService(userID);
-      deleteService.start();
-      deleteService.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
-        @Override
-        public void handle(WorkerStateEvent e) {
-          userService.restart();
-        }
-      });
+      if (alert.getResult().getText().equals("OK")) {
+        deleteService = new DeleteUserService(userID);
+        deleteService.start();
+        deleteService.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
+          @Override
+          public void handle(WorkerStateEvent e) {
+            userService.restart();
+          }
+        });
+      }
     }
   }
 
@@ -78,12 +82,16 @@ public class AccountManagementController {
     userTable.getColumns().setAll(userIDColumn, emailColumn, roleColumn);
     userTable.setPlaceholder(new Label("No users found."));
     userTable.disableProperty().bind(userService.runningProperty());
+    refreshButton.disableProperty().bind(userService.runningProperty());
+    deleteButton.disableProperty()
+        .bind(userService.runningProperty().or(userTable.getSelectionModel().selectedItemProperty().isNull()));
 
     userService.start();
     userService.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
       @Override
       public void handle(WorkerStateEvent e) {
         userTable.setItems(parseUsers());
+        userTable.getColumns().setAll(userIDColumn, emailColumn, roleColumn);
       }
     });
     userService.setOnFailed(new EventHandler<WorkerStateEvent>() {
@@ -111,12 +119,6 @@ public class AccountManagementController {
   }
 
   private ObservableList<UserData> toObservableList(List<UserData> list) {
-    ObservableList<UserData> observableList = FXCollections.observableArrayList();
-
-    for (int i = 0; i < list.size(); i++) {
-      observableList.add(list.get(i));
-    }
-
-    return observableList;
+    return FXCollections.observableArrayList(list);
   }
 }
